Still more rework
