+++ node/pkg/common/chainlock.go
@@ -23,27 +23,27 @@ type MessagePublication struct {
 	Payload          []byte
 }
 
-func (k *MessagePublication) MessageID() []byte {
-	return []byte(k.MessageIDString())
+func (msg *MessagePublication) MessageID() []byte {
+	return []byte(msg.MessageIDString())
 }
 
-func (k *MessagePublication) MessageIDString() string {
-	return fmt.Sprintf("%v/%v/%v", uint16(k.EmitterChain), k.EmitterAddress, k.Sequence)
+func (msg *MessagePublication) MessageIDString() string {
+	return fmt.Sprintf("%v/%v/%v", uint16(msg.EmitterChain), msg.EmitterAddress, msg.Sequence)
 }
 
 const minMsgLength = 88
 
-func (k *MessagePublication) Marshal() ([]byte, error) {
+func (msg *MessagePublication) Marshal() ([]byte, error) {
 	buf := new(bytes.Buffer)
 
-	buf.Write(k.TxHash[:])
-	vaa.MustWrite(buf, binary.BigEndian, uint32(k.Timestamp.Unix()))
-	vaa.MustWrite(buf, binary.BigEndian, k.Nonce)
-	vaa.MustWrite(buf, binary.BigEndian, k.Sequence)
-	vaa.MustWrite(buf, binary.BigEndian, k.ConsistencyLevel)
-	vaa.MustWrite(buf, binary.BigEndian, k.EmitterChain)
-	buf.Write(k.EmitterAddress[:])
-	buf.Write(k.Payload)
+	buf.Write(msg.TxHash[:])
+	vaa.MustWrite(buf, binary.BigEndian, uint32(msg.Timestamp.Unix()))
+	vaa.MustWrite(buf, binary.BigEndian, msg.Nonce)
+	vaa.MustWrite(buf, binary.BigEndian, msg.Sequence)
+	vaa.MustWrite(buf, binary.BigEndian, msg.ConsistencyLevel)
+	vaa.MustWrite(buf, binary.BigEndian, msg.EmitterChain)
+	buf.Write(msg.EmitterAddress[:])
+	buf.Write(msg.Payload)
 
 	return buf.Bytes(), nil
 }
@@ -54,7 +54,7 @@ func UnmarshalMessagePublication(data []byte) (*MessagePublication, error) {
 		return nil, fmt.Errorf("message is too short")
 	}
 
-	k := &MessagePublication{}
+	msg := &MessagePublication{}
 
 	reader := bytes.NewReader(data[:])
 
@@ -62,27 +62,27 @@ func UnmarshalMessagePublication(data []byte) (*MessagePublication, error) {
 	if n, err := reader.Read(txHash[:]); err != nil || n != 32 {
 		return nil, fmt.Errorf("failed to read TxHash [%d]: %w", n, err)
 	}
-	k.TxHash = txHash
+	msg.TxHash = txHash
 
 	unixSeconds := uint32(0)
 	if err := binary.Read(reader, binary.BigEndian, &unixSeconds); err != nil {
 		return nil, fmt.Errorf("failed to read timestamp: %w", err)
 	}
-	k.Timestamp = time.Unix(int64(unixSeconds), 0)
+	msg.Timestamp = time.Unix(int64(unixSeconds), 0)
 
-	if err := binary.Read(reader, binary.BigEndian, &k.Nonce); err != nil {
+	if err := binary.Read(reader, binary.BigEndian, &msg.Nonce); err != nil {
 		return nil, fmt.Errorf("failed to read nonce: %w", err)
 	}
 
-	if err := binary.Read(reader, binary.BigEndian, &k.Sequence); err != nil {
+	if err := binary.Read(reader, binary.BigEndian, &msg.Sequence); err != nil {
 		return nil, fmt.Errorf("failed to read sequence: %w", err)
 	}
 
-	if err := binary.Read(reader, binary.BigEndian, &k.ConsistencyLevel); err != nil {
+	if err := binary.Read(reader, binary.BigEndian, &msg.ConsistencyLevel); err != nil {
 		return nil, fmt.Errorf("failed to read consistency level: %w", err)
 	}
 
-	if err := binary.Read(reader, binary.BigEndian, &k.EmitterChain); err != nil {
+	if err := binary.Read(reader, binary.BigEndian, &msg.EmitterChain); err != nil {
 		return nil, fmt.Errorf("failed to read emitter chain: %w", err)
 	}
 
@@ -90,14 +90,14 @@ func UnmarshalMessagePublication(data []byte) (*MessagePublication, error) {
 	if n, err := reader.Read(emitterAddress[:]); err != nil || n != 32 {
 		return nil, fmt.Errorf("failed to read emitter address [%d]: %w", n, err)
 	}
-	k.EmitterAddress = emitterAddress
+	msg.EmitterAddress = emitterAddress
 
 	payload := make([]byte, vaa.InternalTruncatedPayloadSafetyLimit)
 	n, err := reader.Read(payload)
 	if err != nil || n == 0 {
 		return nil, fmt.Errorf("failed to read payload [%d]: %w", n, err)
 	}
-	k.Payload = payload[:n]
+	msg.Payload = payload[:n]
 
-	return k, nil
+	return msg, nil
 }
